commit for validation and logging

+++ BookingDemo/src/main/java/com/example/demo/BookingController.java
@@ -1,7 +1,12 @@
 package com.example.demo;
 
 import java.util.List;
+import java.util.Optional;
 
+import javax.validation.Valid;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -10,9 +15,14 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.example.demo.exception.FlightNotFoundException;
+
 @RestController
 @RequestMapping(value = "/bookings")
 public class BookingController {
+	
+	private static final Logger LOGGER = LoggerFactory
+			.getLogger(BookingController.class);
 
 	private BookingRepository bookingRepository;
 	
@@ -26,26 +36,40 @@ public class BookingController {
 	@GetMapping(path = "/all")
 	public List<FlightBooking> getAllBookings()
 	{
+		LOGGER.info("getAllBookings- {}", "getAllBookings start function");          
+		LOGGER.info("getAllBookings- {}", "getAllBookings end function");
 		return bookingRepository.findAll();
 	}
 	
 	@GetMapping(path = "/affordable/{price}")
-	public List<FlightBooking> getAffordableBooking(@PathVariable double price)
+	public List<FlightBooking> getAffordableBooking(@Valid @PathVariable double price)
 	{
+		LOGGER.info("getAffordableBooking- {}", "getAffordableBooking start function");          
+		LOGGER.info("getAffordableBooking- {}", "getAffordableBooking end function");
 		return bookingRepository.findByPriceLessThan(price)	;
 	}
 	
 	@PostMapping(path = "/create")
-	public List<FlightBooking> createBookings(@RequestBody FlightBooking flightBooking )
+	public List<FlightBooking> createBookings(@Valid @RequestBody FlightBooking flightBooking )
 	{
+		LOGGER.info("createBookings- {}", "createBookings start function");
 		bookingRepository.save(flightBooking);
+		LOGGER.info("createBookings- {}", "createBookings end function");
 		return bookingRepository.findAll();
 	}
 	
 	@PostMapping(path = "/delete/{id}")
-	public List<FlightBooking> deleteBookings(@PathVariable long id)
+	public List<FlightBooking> deleteBookings(@Valid @PathVariable long id)
 	{
+		LOGGER.info("deleteBookings- {}", "deleteBookings start function");
+		Optional<FlightBooking> flightBooking = bookingRepository.findById(id);
+		if(!flightBooking.isPresent())
+		{
+			LOGGER.error("error scenario- {}", "No booking present to delete");
+			throw new FlightNotFoundException("No booking present");
+		}
 		bookingRepository.deleteById(id);
+		LOGGER.info("deleteBookings- {}", "deleteBookings end function");
 		return bookingRepository.findAll();
 	}
 }
